New Features: for batch_prepare_tiltseries, warn if no ExposureDose is specified manually or in mdoc. Also allows batch_prepare_tiltseries if no subframes are found. for reconstruct, should now tolerate more binning levels and be agnostic how tilt axis and image xy relate. Can skip dose filtration if no ExposureDose is given in mdoc. Bug Fix: For trimming a reconstruction, read out dimension from the full_rec file to avoid binning imprecisions leading to failed trim command. Added test files for different cameras & calibrations to test folder. Started writing automated test script.

+++ tomotools/commands/tomography.py
@@ -72,6 +72,7 @@ def batch_prepare_tiltseries(splitsum, mcbin, reorder, frames, gainref, group, g
     The last argument is the output dir. It will be created if it doesn't exist.
     """
     # If input_files is a directory, check for PACEtomo files by regex _tgts.txt, save all roots
+    # TODO: We might be able to skip this as there's now automated detection of .mrc files which are not tilt series
     wd = os.getcwd()
 
     if any(path.isdir(input_file) for input_file in input_files):
@@ -143,8 +144,10 @@ def batch_prepare_tiltseries(splitsum, mcbin, reorder, frames, gainref, group, g
             )
             if exposuredose is not None:
                 section['ExposureDose'] = exposuredose
-            # TODO: warning if ExposureDose = 0 and not set
-            #print(f'SubFramePath field: {section.get("SubFramePath", "")}')
+                        
+        if any(section['ExposureDose'] == 0 for section in mdoc['sections']) and exposuredose is None:
+            print(f'{input_file} has no ExposureDose set. This might lead to problems down the road!')   
+        
         # Check if all subframes were found
         subframes = [frame_utils.SubFrame(section['SubFramePath'],section['TiltAngle']) for section in mdoc['sections']]
         if all(subframe.files_exist(is_split=False) for subframe in subframes):
@@ -171,11 +174,23 @@ def batch_prepare_tiltseries(splitsum, mcbin, reorder, frames, gainref, group, g
             shutil.rmtree(frames_corrected_dir)
             print(f'Successfully created {path.basename(stack)}')
         else:
-            print(f'No subframes were found for {input_file}, will continue without MotionCor2')
-            # TODO: add implementation for non-subframe TS w/ and w/o reorder!            
-            print('Not implemented yet, skipping')
-            continue
-            # raise NotImplementedError('Sorry, non-motioncor2 path is not implemented yet')
+            if reorder:    
+                subprocess.run(['newstack',
+                                '-reorder', 1,
+                                '-mdoc',
+                                '-in', input_file,
+                                '-mdoc',
+                                '-ou', f'{output_dir}/{input_file}'])  
+                print(f'No subframes were found for {input_file}, ran newstack -reorder')
+
+            else:
+                subprocess.run(['cp',
+                                input_file,
+                                output_dir])
+                subprocess.run(['cp',
+                                f'{input_file}.mdoc',
+                                output_dir])
+                print(f'No subframes found and no reorder requested. Just copying {input_file} to output directory.')
 
 @click.command()
 @click.option('--move', is_flag=True, help="Move files into a subdirectory")
@@ -227,7 +242,6 @@ def reconstruct(move, local, extra_thickness, bin, sirt, keep_ali_stack, previou
                 mdoc_file = new_mdoc_file
 
         # Run newstack to exclude tilts
-        # TODO: Consider already binning here before AreTomo?
         rootname = splitext(tiltseries)[0]
 
         if excludetilts is not None:
@@ -242,13 +256,28 @@ def reconstruct(move, local, extra_thickness, bin, sirt, keep_ali_stack, previou
             tiltseries = exclude_file
             mdoc_file = f'{tiltseries}.mdoc'
 
+        # Read full stack size from mdoc header        
         mdoc = mdocfile.read(mdoc_file)
+        
+        full_dimensions = mdoc['ImageSize']
 
-        # To account for different sensor sizes, read from mdoc file.
+        # Read pixelsize and image dimensions from brief header in case the stack was binned.
+        header = subprocess.run(['header', '-brief',tiltseries], capture_output=True, text=True).stdout
+        header = header.strip().split()
+        
+        stack_dimensions = [int(header[header.index('Dimensions:')+1]),int(header[header.index('Dimensions:')+2])]
+        pix_xy = [float(header[header.index('size:')+1]),float(header[header.index('size:')+2])]       
+        
+        # Get tilt axis from full header
+        header = subprocess.run(['header',tiltseries], capture_output=True, text=True).stdout
+        header = header.strip()
+        
+        tilt_axis = float(header[header.index("Tilt axis angle =") + len("Tilt axis angle =") + 1:header.index(", binning")])
+        
         # Define default thickness as function of pixel size -> always reconstruct 1 um for tomopitch. 
-        thickness = str(round(10000 / mdoc['PixelSpacing']))
-        full_x, full_y = mdoc['ImageSize']
-        patch_x, patch_y = [str(round(full_x/1000)), str(round(full_y/1000))]
+        thickness = str(round(10000 / pix_xy[0]))
+        
+        patch_x, patch_y = [str(round(full_dimensions[0]/1000)), str(round(full_dimensions[1]/1000))]
 
         # Generate MRC filenames
         # TODO: Make a class?  
@@ -266,7 +295,6 @@ def reconstruct(move, local, extra_thickness, bin, sirt, keep_ali_stack, previou
         aln_file = f'{ali_rootname}.aln'
 
         # Run AreTomo 
-        # TODO: Read in Tiltaxis?
         if previous is None:
             subprocess.run(['extracttilts', tiltseries, tlt_file],
                            stdout=subprocess.DEVNULL)
@@ -287,19 +315,30 @@ def reconstruct(move, local, extra_thickness, bin, sirt, keep_ali_stack, previou
 
         print(f'Done aligning {tiltseries} with AreTomo.')
 
-        # TODO: Discuss whether to directly use the value from the mdoc here
-        pix_size = subprocess.run(['header', '-PixelSize', tiltseries], capture_output=True, text=True).stdout
-        pix_size = ",".join(pix_size.strip().split())
-        subprocess.run(['alterheader', '-PixelSize', pix_size, ali_file],
+        # Apply pixel size from input stack to aligned stack
+        pix_xyz = subprocess.run(['header', '-PixelSize', tiltseries], capture_output=True, text=True).stdout
+        pix_xyz = ",".join(pix_xyz.strip().split())
+        subprocess.run(['alterheader', '-PixelSize', pix_xyz, ali_file],
                        stdout=subprocess.DEVNULL)
 
-        # Dose filtration. Only ExposureDose needs to be entered, as PriorRecordDose is deduced by mtffilter based on the DateTime entry, see mtffilter -help, section "-dtype"
-        subprocess.run(['mtffilter', '-dtype', '4', '-dfile', mdoc_file, ali_file, ali_file_mtf],
-                       stdout=subprocess.DEVNULL)
-        if not keep_ali_stack:
-            os.remove(ali_file)
+        # If exposuredose is set, do dose filtration. Only ExposureDose needs to be entered, as PriorRecordDose is deduced by mtffilter based on the DateTime entry, see mtffilter -help, section "-dtype"
+        if any(section['ExposureDose'] == 0 for section in mdoc['sections']):
+            ali_file_mtf = ali_file
+            print(f'{tiltseries} has no ExposureDose set. Skipping dose-filtration.')
+            
+        else:
+            subprocess.run(['mtffilter', '-dtype', '4', '-dfile', mdoc_file, ali_file, ali_file_mtf],
+                           stdout=subprocess.DEVNULL)
+            if not keep_ali_stack:
+                os.remove(ali_file)
+            print(f'Done dose-filtering {tiltseries}.')
 
-        print(f'Done dose-filtering {tiltseries}.')
+        # Set full_reconstruction_x and full_reconstruction_y based on TiltAxis property. Imod convention: tilt axis is x axis. 
+        if tilt_axis > 45 and tilt_axis < 135:
+            full_reconstruction_x = stack_dimensions[1]
+            full_reconstruction_y = stack_dimensions[0]
+        else:
+            [full_reconstruction_x,full_reconstrucion_y] = stack_dimensions
 
         # Tomo pitch
         if previous is None:
@@ -316,7 +355,7 @@ def reconstruct(move, local, extra_thickness, bin, sirt, keep_ali_stack, previou
                            '-SCALE', '0.0,0.05',
                            '-PERPENDICULAR',
                            '-MODE', '2',
-                           '-FULLIMAGE', f'{full_y} {full_x}',
+                           '-FULLIMAGE', f'{full_reconstruction_x} {full_reconstruction_y}',
                            '-SUBSETSTART', '0,0',
                            '-AdjustOrigin',
                            '-ActionIfGPUFails', '1,2',
@@ -343,9 +382,8 @@ def reconstruct(move, local, extra_thickness, bin, sirt, keep_ali_stack, previou
                 x_axis_tilt = '0'
                 tomopitch_z = '0'
                 z_shift = '0'
-                thickness = str(thickness)
+                thickness = str(round(6000 / pix_xy[0])+extra_thickness)
                 print(f'{tiltseries}: findsection failed, using default values {tomopitch_z}, thickness {thickness}.')
-
 
             else:
                 # Else, get tomopitch
@@ -359,14 +397,14 @@ def reconstruct(move, local, extra_thickness, bin, sirt, keep_ali_stack, previou
                     x_axis_tilt = '0'
                     tomopitch_z = '0'
                     z_shift = '0'
-                    thickness = str(thickness)
+                    thickness = str(round(6000 / pix_xy[0])+extra_thickness)
                     print(f'{tiltseries}: tomopitch failed, using default values {tomopitch_z}, thickness {thickness}.')
                 else:
                     x_axis_tilt = tomopitch[-3].split()[-1]
                     tomopitch_z = tomopitch[-1].split(';')
                     z_shift = tomopitch_z[0].split()[-1]
-                    thickness = tomopitch_z[1].split()[-1]
-                    print(f'{tiltseries}: Succesfully estimated tomopitch {tomopitch_z} and thickness {thickness}.')
+                    thickness = str(int(tomopitch_z[1].split()[-1])+extra_thickness)
+                    print(f'{tiltseries}: Succesfully estimated tomopitch {x_axis_tilt} and thickness {thickness}.')
 
         # Final reconstruction
         if bin == 1:
@@ -392,7 +430,7 @@ def reconstruct(move, local, extra_thickness, bin, sirt, keep_ali_stack, previou
                         '-SCALE', '0.0,0.05',
                         '-PERPENDICULAR',
                         '-MODE', '2',
-                        '-FULLIMAGE', f'{full_y} {full_x}',
+                        '-FULLIMAGE', f'{full_reconstruction_x} {full_reconstruction_y}',
                         '-SUBSETSTART', '0,0',
                         '-AdjustOrigin',
                         '-ActionIfGPUFails', '1,2',
@@ -405,22 +443,28 @@ def reconstruct(move, local, extra_thickness, bin, sirt, keep_ali_stack, previou
 
         print(f'{tiltseries}: Finished reconstruction.')
 
-        # Trim
-        # TODO: Consider case where image is not rotated 90deg in relation to mdoc (TiltAxis property?)
-        thickness = int(thickness)
-        subprocess.run(['trimvol',
-                        '-x', f'1,{full_y / bin:.0f}',
-                        '-y', f'1,{full_x / bin:.0f}',
-                        '-z', f'1,{thickness / bin:.0f}',
-                        '-sx', f'1,{full_y / bin:.0f}',
-                        '-sy', f'1,{full_x / bin:.0f}',
-                        '-sz', f'{thickness / bin / 3:.0f},{thickness / bin * 2 / 3:.0f}',
+        # Trim: Read in dimensions of full_rec (as XZY) to avoid rounding differences due to binning
+        header_rec = subprocess.run(['header', '-brief',full_rec_file], capture_output=True, text=True).stdout
+        header_rec = header_rec.strip().split()
+        
+        full_rec_dimensions = [header_rec[header_rec.index('Dimensions:')+1],header_rec[header_rec.index('Dimensions:')+2],header_rec[header_rec.index('Dimensions:')+3]]
+
+        tr = subprocess.run(['trimvol',
+                        '-x', f'1,{full_rec_dimensions[0]}',
+                        '-y', f'1,{full_rec_dimensions[2]}',
+                        '-z', f'1,{full_rec_dimensions[1]}',
+                        '-sx', f'1,{full_rec_dimensions[0]}',
+                        '-sy', f'1,{full_rec_dimensions[2]}',
+                        '-sz', f'{int(full_rec_dimensions[1]) / 3:.0f},{int(full_rec_dimensions[1]) * 2 / 3:.0f}',
                         '-f', '-rx',
                         full_rec_file, rec_file],
                         stdout=subprocess.DEVNULL)
 
-        print(f'{tiltseries}: Finished trimming.')
-
-        os.remove(full_rec_file)
+        if tr.returncode != 0:
+            print(f'{tiltseries}: Trimming failed, keeping full_rec file.')        
+        else:
+            print(f'{tiltseries}: Finished trimming.')
+            os.remove(full_rec_file)
+            
         if path.isfile('mask3000.mrc'):
             os.remove('mask3000.mrc')
